* 1.2.0.redhat-6-2-x-patch:
  [ENTESB-4367] Carefuly handle LF/CRLF under Windows
  [jboss-fuse-6.2.x 076] Updating perfectus-build number to new build
  Revert "track down intermittent failure of ServiceFactoryTest. Root cause was partial sequential ephemeral node creation leading to duplicate state and conditional guarantee delete rather than letting curator take care of reconnect. Fix is to use hidden uuid nodestate field on znode creation such that duplicates can be suppressed"
  [ENTESB-4431] ensure zk-group json parser ignores unknown fields
  [jboss-fuse-6.2.x 075] Updating perfectus-build number to new build
  ENTESB-4398 - update to commons-collections 3.2.1.redhat-5
  [jboss-fuse-6.2.x 073] Updating perfectus-build number to new build
  [ENTESB-4387] Fixing patch:fabric-synchronize description and output of the command
  [ENTESB-4362] Upgrading to jgit 4.1 with better pack management
  [ENTESB-4415] Upgrading features with some order hints
  [ENTESB-4422] Fix construction of URLs from file paths on Windows
  [ENTESB-4415] Upgrade to jackson-module-scala 2.1.5_3 with better scala import range
  [jboss-fuse-6.2.x 072] Updating perfectus-build number to new build
  [ENTESB-4393] NPE when custom resolving conflict and there is no "common base"
  [ENTESB-4374] Do not treat patch branches as versions
  [ENTMQ-1314] Fix fabric:create wrt patching in AMQ distro
  track down intermittent failure of ServiceFactoryTest. Root cause was partial sequential ephemeral node creation leading to duplicate state and conditional guarantee delete rather than letting curator take care of reconnect. Fix is to use hidden uuid nodestate field on znode creation such that duplicates can be suppressed

+++ patch/patch-management/src/main/java/io/fabric8/patch/management/impl/GitPatchManagementServiceImpl.java
@@ -515,6 +515,7 @@ public class GitPatchManagementServiceImpl implements PatchManagement, GitPatchM
                     Activator.log2(LogService.LOG_DEBUG, String.format("Uploaded %d/%d", count, artifacts.size()));
                 }
                 String relativeName = Utils.relative(Utils.getSystemRepository(karafHome, bundleContext), f.getCanonicalFile());
+                relativeName = relativeName.replace('\\', '/');
                 URL uploadUrl = uploadAddress.resolve(relativeName).toURL();
                 URLConnection con = uploadUrl.openConnection();
                 callback.doWithUrlConnection(con);
@@ -1289,10 +1290,13 @@ public class GitPatchManagementServiceImpl implements PatchManagement, GitPatchM
                     // when doing custom resolution, we prefer user change
                     File base = null, first = null, second = null;
                     try {
-                        base = new File(fork.getRepository().getWorkTree(), entry.getKey() + ".1");
-                        ObjectLoader loader = objectReader.open(entry.getValue()[1]);
-                        try (FileOutputStream fos = new FileOutputStream(base)) {
-                            loader.copyTo(fos);
+                        ObjectLoader loader = null;
+                        if (entry.getValue()[1] != null) {
+                            base = new File(fork.getRepository().getWorkTree(), entry.getKey() + ".1");
+                            loader = objectReader.open(entry.getValue()[1]);
+                            try (FileOutputStream fos = new FileOutputStream(base)) {
+                                loader.copyTo(fos);
+                            }
                         }
 
                         // if preferNew (P patch) then first will be change from patch
@@ -1305,7 +1309,7 @@ public class GitPatchManagementServiceImpl implements PatchManagement, GitPatchM
                         second = new File(fork.getRepository().getWorkTree(), entry.getKey() + ".3");
                         loader = objectReader.open(entry.getValue()[preferNew ? 0 : 2]);
                         try (FileOutputStream fos = new FileOutputStream(second)) {
-                            loader.copyTo(new FileOutputStream(second));
+                            loader.copyTo(fos);
                         }
 
                         // resolvers treat patch change as less important - user lines overwrite patch lines
@@ -1620,7 +1624,7 @@ public class GitPatchManagementServiceImpl implements PatchManagement, GitPatchM
                         .setTagOpt(TagOpt.FETCH_TAGS)
                         .call();
 
-                if (env == EnvType.FABRIC_FUSE) {
+                if (env == EnvType.FABRIC_FUSE || env == EnvType.FABRIC_AMQ) {
                     // I think it's enough to compare the main HEADs. if there are no remote branches
                     // for patch management or their HEADs are the same, we are the MAIN container that performs
                     // git patch management
@@ -2668,14 +2672,16 @@ public class GitPatchManagementServiceImpl implements PatchManagement, GitPatchM
                 return VERSION_PATTERN.matcher(name).matches();
             }
         });
-        for (File anotherPatchManagementBundle : deployedPatchManagementBundles) {
-            Matcher matcher = VERSION_PATTERN.matcher(anotherPatchManagementBundle.getName());
-            matcher.find();
-            String version = matcher.group(1);
-            Version deployedVersion = new Version(version);
-            if (ourVersion.compareTo(deployedVersion) >= 0) {
-                Activator.log(LogService.LOG_INFO, "Deleting " + anotherPatchManagementBundle);
-                FileUtils.deleteQuietly(anotherPatchManagementBundle);
+        if (deployedPatchManagementBundles != null) {
+            for (File anotherPatchManagementBundle : deployedPatchManagementBundles) {
+                Matcher matcher = VERSION_PATTERN.matcher(anotherPatchManagementBundle.getName());
+                matcher.find();
+                String version = matcher.group(1);
+                Version deployedVersion = new Version(version);
+                if (ourVersion.compareTo(deployedVersion) >= 0) {
+                    Activator.log(LogService.LOG_INFO, "Deleting " + anotherPatchManagementBundle);
+                    FileUtils.deleteQuietly(anotherPatchManagementBundle);
+                }
             }
         }
     }
